feat: usage/cache logic to store usage metric
Adding new usage and cache logic to account-usage package that will
power usage capping and reporting.

Metric count will be store in Redis and revalidated periodically in the
background. The objective is to have quick capping mechanism  since in
most case it must be done inline, while treating the data in Redis as
cached data that will be revalidated from the source of truth (orb, db) regularly