fix(runner-sdk): better handle record schema compilation
when validating saved records we compile and cache the model schema
if the compilation fails for some reason the compilation is attempted
again on all batchSave (and keeps failing).
Ajv compilation seems to hold some references/data in memory and since
we use a static ajv instance the memory is never released and piling up.

This commit ensures we also cache compilation failure and not re-attempt
to compile a failing schema.
It doesn't solve the static ajv instance holding on memory but it would
take 1000s of different models/versions to run into some kind of memory
leak. To keep in mind though.