fix(record count): do not count records from deleted connections
Record counts table entries are not deleted when a connection is soft
deleted (only when hard deleted) which means that records are still
counted until the connection is hard deleted.
Since we delete connections from different places in the codebase it would require a
big refactoring to delete the record counts table when connections is
soft deleted. Main pain point is connections are soft-deleted when an
integration is soft-deleted, deep inside the shared package, which
hasn't any knowledge of the records packages :(
Instead I decided to account for the deleted connections when we report
the record counts. The logic was already quite cumbersome and wasn't
supporting pagination.