fix(usage): account for deleted connections when revalidating records
I fixed the issue where we were not accounting for deleted connections
when reporting records count but I forgot the records count data was
also used directly in usage revalidation.
This commit remove the records.metrics function that cannot be used
direclty and ensure records for deleted connections are not counted.
Because records and connections don't live in the same db, joins aren't
possible and we have to do some cumbersome gymnastic in the code.
It is not super scalable but since we are dealing with a very small
number of environments it is still very manageable in term of
performance