chore(usage): replace rate limiter by queue when throttling
Instead of failing when fetching usage hit the rate limit, queue the
requests to be executed in next time interval.
It could eventually fail if we hit the queue max size but since we are
gonna request a limited number of time interval and aggresivelly cache
the data it is only likely to happen if lots of customers hit their
billing page at the same time.
It would have the effect to make loading the usage data slower which I
believe it would be a better user experience that failing and having to
manage retries, etc...