fix(runners): keep track of pending actions
Idling logic depends on keeping track of ongoing task execution
We used to not keep track of actions, relying on the fact that actions
are usually short running and runner idling time was 24h
We recently dropped the idling time to 30mins which made it possible for
runners doing only actions to always idle after 30mins, requiring a cold
start.
This commit modifies the tasks tracking logic on runner, tracking all
pending tasks (including actions). It is gonna increase the memory
consumption of actions heavy runner but will make the idling logic less
brittle and timing dependent