chore(webhooks): add circuit breaker
Adding a circuit breaker mechanism to fail quickly when a customer
webhook endpoint is unresponsive (timeouts/errors). The goal is to
protect Nango services when a customer webhook endpoint is down.
When requests for a given endpoint fails beyond the failure threshold
Nango will skip making the requests and fail the webhook delivery
quickly. After a cooldown period Nango will let at least one request go
through. If recovery requests are successful, the circuit breaker will be
closed and delivering the webhook will be attempted again as normal for
the url. If recovery requests failed, delivery of webhooks for this url
will be stopped again until the cooldown period ends.
This commit is also lowering the timeout of the webhook delivery
requests to 20s (from 60s) and the number of retry attempt to 2 (previously 7)

This is not introducing any queuing mechanism and a large amount of
webhook deliveries might still be attempted before the circuit breaker open. (if for example the customer endpoint is slow to respond). It would still provide some protection by stopping outgoing requests once the 20s timeoutis elapsed and won't allow a large amount of requests until the endpoint responds with success again). It could however introduce a seesaw behavior if the customer endpoints slow down under load and recover quickly.