chore(records): syncId not necessary to delete records
record.syncId is a legacy field. I don't remember why it was ported from
the old data model but I imagine it was required for the transition
It is currently only inserted but also used for deletion. However since
a given connection/model is already unique (variant is part of the model
name), filtering deletion by syncId is unecessary.