feat(records): replace DELETE endpoint by PATCH /records/prune
I misunderstand the requirements and thought pruning was linked to
records deletion but records pruning is actually its own operation,
independant to the concept of deleting a record
This commit is removing the DELETE /records endpoint and replace it by
PATCH /records/prune
I have kept the internal deleteRecords function in order to take
advantage of the shared logic (pagination, etc...). For now soft delete,
hard delete, prune are disjoint options of the deleteRecords function
but could make it a bitwise paramter if we need to combine them in the
future.