chore: fixedSizedMap utilt
Implement a FixedSizeMap util class to use as bounded local cache.
We are using `Map` in a few places as local caches. Even though their size
under normal operations is likely to be very reasonable it would be
safer to have limits in place to prevent high memory usage under
extraordinary circumstances.